import product

+++ src/main/java/phone/shop/repository/ModelRepository.java
@@ -4,5 +4,5 @@ import phone.shop.model.Model;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface ModelRepository extends JpaRepository<Model,Integer>, JpaSpecificationExecutor<Model> {
+public interface ModelRepository extends JpaRepository<Model,Long>, JpaSpecificationExecutor<Model> {
 }
